In general, refactoring is done but LLM still doesn't use RAG

+++ ingestor/services/knowledge.py
@@ -151,4 +151,4 @@ class KnowledgePort:
         if norm1 == 0 or norm2 == 0:
             return 0.0
         
-        return dot_product / (norm1 * norm2)
+        return float(dot_product / (norm1 * norm2))
